Moved weather code to services, don't use though because very slow
