+++ job/services.py
@@ -8,8 +8,7 @@ from reportlab.lib.utils import ImageReader
 from reportlab.lib.styles import ParagraphStyle
 from reportlab.lib.enums import TA_RIGHT
 from io import BytesIO
-import re
-import os
+import os, re, urllib, json
 
 def get_all_jobs():
     jobs = Job.objects.all().order_by('date_of_job')
@@ -40,6 +39,9 @@ materialStyle = ParagraphStyle(
 
 # add data to pdf
 def add_data(job):
+    """
+    Add all the data from the job to the pdf.
+    """
     document = []
     document.append(get_image('./general/static/general/img/logosmall.jpg', width=8*cm))
     document.append(Spacer(1, 30))
@@ -164,3 +166,102 @@ def get_image(path, width):
     iw, ih = img.getSize()
     aspect = ih / float(iw)
     return Image(path, width=width, height=(width * aspect))
+
+def get_coords(location):
+    location = location.replace(" ", '+')
+    api_key = os.environ.get("OPENCAGE_KEY")
+    query_url = 'https://api.opencagedata.com/geocode/v1/json?q=' + location + '&key=' + api_key 
+
+    inp = urllib.request.urlopen(query_url)
+    coords = json.load(inp)['results'][0]['geometry']
+
+    return coords
+
+def format_coords(coordinates):
+    coordinates['lat'] = round(coordinates['lat'], 2)
+    coordinates['lng'] = round(coordinates['lng'], 2)
+
+    lat = str(coordinates['lat'])
+    if lat[::-1].find('.') == 1:
+        lat += '0'
+    lat = lat.replace('.', 'd').replace('-', 'n')
+
+    lng = str(coordinates['lng'])
+    if lng[::-1].find('.') == 1:
+        lng += '0'
+    lng = lng.replace('.', 'd').replace('-', 'n')
+
+    return lat + lng
+
+
+def get_widget_code(coords, location, lang):
+    A_TAG = """
+    <a 
+        class="weatherwidget-io big-widget" 
+        href="https://forecast7.com/_LANG_/_COORDS_/_LOCATION_/" 
+        data-label_1="_LOCATIONPRETTY_" data-label_2="WEATHER" 
+        data-font="Roboto" data-theme="pure"
+    >
+        _LOCATIONPRETTY_ {{_("WEATHER")}}
+    </a>
+    """
+    
+    coords = format_coords(coords)
+
+    tag = A_TAG.replace("_COORDS_", coords)
+    tag = tag.replace("_LOCATIONPRETTY_", location)
+    tag = tag.replace("_LOCATION_", location)
+    tag = tag.replace("_LANG_", lang)
+    url = "https://forecast7.com/_LANG_/_COORDS_/_LOCATION_/"
+    url = url.replace("_COORDS_", coords).replace("_LOCATION_", location).replace("_LANG_", lang)
+
+    return tag, url
+
+def get_widget_html(pretty_name, lang, units=None, formated_coords=None):
+    if formated_coords is None:
+        coords = get_coords(pretty_name)
+        formated_coords = format_coords(coords)
+
+    tag_code, url = get_widget_code(formated_coords, pretty_name, lang)
+    if not is_url_ok(url):
+        tag_code, url = fix_url(coords, pretty_name, lang)
+
+    return tag_code + SCRIPT
+
+
+def is_url_ok(url):
+    try:
+        req = urllib.request.Request(
+            url, headers={'User-Agent': "Magic Browser"})
+        urllib.request.urlopen(req)
+        return True
+
+    except urllib.error.HTTPError as e:
+        print(e)
+        return False
+
+
+def fix_url(coords, pretty_name, lang):
+    tolerance = 5
+    if tolerance:
+        for i in range(-tolerance, tolerance + 1):
+            nc = coords['lat'] + i / 100
+            for j in range(-tolerance, tolerance + 1):
+                ncl = coords['lng'] + j / 100
+                formated_coords = {'lat': nc, 'lng': ncl}
+
+                tag_code, url = get_widget_code(formated_coords, pretty_name, lang)
+                if is_url_ok(url):
+                    return tag_code, url
+
+
+def get_widget_html(pretty_name, lang, units=None, formated_coords=None):
+    if formated_coords is None:
+        coords = get_coords(pretty_name)
+        formated_coords = format_coords(coords)
+    tag_code, url = get_widget_code(formated_coords, pretty_name, lang)
+    
+    if not is_url_ok(url):
+        tag_code, url = fix_url(coords, pretty_name, lang)
+
+    return tag_code + SCRIPT
