+++ job/views.py
@@ -3,7 +3,7 @@ from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.views.generic import ListView, DetailView, UpdateView, DeleteView
 from job.models import Job, JobForm
-from job.services import get_all_jobs, get_single_job, create_pdf
+from job.services import get_all_jobs, get_single_job, create_pdf, get_coords, get_widget_code, fix_url, is_url_ok
 from quote.services import get_single_quote
 from django.db.models import Q
 from django.urls import reverse_lazy
@@ -11,7 +11,7 @@ from django.core.mail import send_mail, EmailMessage
 from django.views.decorators.csrf import csrf_exempt
 from django.http import JsonResponse, HttpResponseRedirect
 from smtplib import SMTPException
-import os, urllib, json
+import os
 
 
 @csrf_exempt
@@ -69,106 +69,6 @@ class JobListView(LoginRequiredMixin, ListView):
         
         return jobs
 
-def get_coords(location):
-    location = location.replace(" ", '+')
-    api_key = os.environ.get("OPENCAGE_KEY")
-    query_url = 'https://api.opencagedata.com/geocode/v1/json?q=' + location + '&key=' + api_key 
-
-    inp = urllib.request.urlopen(query_url)
-    coords = json.load(inp)['results'][0]['geometry']
-
-    return coords
-
-def format_coords(coordinates):
-    coordinates['lat'] = round(coordinates['lat'], 2)
-    coordinates['lng'] = round(coordinates['lng'], 2)
-
-    lat = str(coordinates['lat'])
-    if lat[::-1].find('.') == 1:
-        lat += '0'
-    lat = lat.replace('.', 'd').replace('-', 'n')
-
-    lng = str(coordinates['lng'])
-    if lng[::-1].find('.') == 1:
-        lng += '0'
-    lng = lng.replace('.', 'd').replace('-', 'n')
-
-    return lat + lng
-
-
-def get_widget_code(coords, location, lang):
-    A_TAG = """
-    <a 
-        class="weatherwidget-io big-widget" 
-        href="https://forecast7.com/_LANG_/_COORDS_/_LOCATION_/" 
-        data-label_1="_LOCATIONPRETTY_" data-label_2="WEATHER" 
-        data-font="Roboto" data-theme="pure"
-    >
-        _LOCATIONPRETTY_ {{_("WEATHER")}}
-    </a>
-    """
-    
-    coords = format_coords(coords)
-
-    tag = A_TAG.replace("_COORDS_", coords)
-    tag = tag.replace("_LOCATIONPRETTY_", location)
-    tag = tag.replace("_LOCATION_", location)
-    tag = tag.replace("_LANG_", lang)
-    url = "https://forecast7.com/_LANG_/_COORDS_/_LOCATION_/"
-    url = url.replace("_COORDS_", coords).replace("_LOCATION_", location).replace("_LANG_", lang)
-
-    return tag, url
-
-def get_widget_html(pretty_name, lang, units=None, formated_coords=None):
-    if formated_coords is None:
-        coords = get_coords(pretty_name)
-        formated_coords = format_coords(coords)
-
-    tag_code, url = get_widget_code(formated_coords, pretty_name, lang)
-    if not is_url_ok(url):
-        tag_code, url = fix_url(coords, pretty_name, lang)
-
-    return tag_code + SCRIPT
-
-
-def is_url_ok(url):
-    try:
-        req = urllib.request.Request(
-            url, headers={'User-Agent': "Magic Browser"})
-        urllib.request.urlopen(req)
-        return True
-
-    except urllib.error.HTTPError as e:
-        print(e)
-        return False
-
-
-def fix_url(coords, pretty_name, lang):
-    tolerance = 5
-    if tolerance:
-        for i in range(-tolerance, tolerance + 1):
-            nc = coords['lat'] + i / 100
-            for j in range(-tolerance, tolerance + 1):
-                ncl = coords['lng'] + j / 100
-                formated_coords = {'lat': nc, 'lng': ncl}
-
-                tag_code, url = get_widget_code(formated_coords, pretty_name, lang)
-                if is_url_ok(url):
-                    return tag_code, url
-
-
-def get_widget_html(pretty_name, lang, units=None, formated_coords=None):
-    if formated_coords is None:
-        coords = get_coords(pretty_name)
-        formated_coords = format_coords(coords)
-    tag_code, url = get_widget_code(formated_coords, pretty_name, lang)
-    
-    if not is_url_ok(url):
-        tag_code, url = fix_url(coords, pretty_name, lang)
-
-    return tag_code + SCRIPT
-
-
 
 class JobDetailView(LoginRequiredMixin, DetailView):
     model = Job
@@ -183,22 +83,14 @@ class JobDetailView(LoginRequiredMixin, DetailView):
         job.city = address[1].lstrip().lower()
         job.post_code = address[2].lstrip()
 
-        print(get_coords(job.city))
-
-        # print(
-        #     get_widget_code(get_coords(job.city), job.city, 'en')
-        # )
-
-        print("below")
         tag_code, url = get_widget_code(get_coords(job.city), job.city, 'en')
         if not is_url_ok(url):
-            print("below 1.1")
             tag_code, url = fix_url(get_coords(job.city), job.city, 'en')
         
         print("below 2")
         print(tag_code)
-        print("below 3")
         print(url)
+        
         return job
     
     def render_to_response(self, context, **response_kwargs):
